Must Question ( Shifting concept ( Hash ) Check )
Link: https://codeforces.com/contest/1365/problem/C
Approach
Since both arrays are a permutation, each element in a would match with its corresponding equal element in b only for one of the shifts. For example, if a is {2,3,1} and b is {3,1,2}, the number 3 in a would match with the number 3 in b only if one right cyclic shift is performed. So for each element in 'a' we can find the number of right cyclic shifts after which it would match with its corresponding equal element in 'b'. If ai = bj, then ai would match with bj after k=j−i -1 right cyclic shifts. If j−i<0, then ai would with bj after n−j+i shifts.
Now for each shift, we can find the number of matching pairs and take the maximum.